// webui/src/app/models/asset-monthly-targets.model.ts
export interface AssetMonthltTargetsModel {
    
    id: number;
    assetType: any;
    scheduleType: any;
    totalPopulation: any;
    year: any;
    targetApr: any;
    targetMay: any;
    targetJune: any;
    targetJuly: any;
    targetAug: any;
    targetSep: any;
    targetOct: any;
    targetNov: any;
    targetDec: any;
    targetJan: any;
    targetFeb: any;
    targetMar: any;    
}